Moved collection registration to command & added caching for page builder registration
