--- config/statamic-page-builder-kit.php
@@ -1,15 +1,0 @@
-<?php
-
-return [
-    /*
-    |--------------------------------------------------------------------------
-    | Boot Collections
-    |--------------------------------------------------------------------------
-    |
-    | When enabled, the package will automatically create and configure the
-    | pages collection with the default structure. Set to false if you want
-    | to manage the collection configuration manually.
-    |
-    */
-    'boot_collections' => true,
-];
